Move the dropdown to the reportHeader component
